Refactor serializers to separate files
